fuctional-testing/requirements.txt
GitPython==3.1.40
pytest==7.4.0
